Add Athena querying utility class
Why these changes are being introduced:

Querying Athena for S3 Inventory is an important part of the
AIP validation.  This standardizes how Athena is queried, using
the library PyAthena as a DB API driver for Athena, and
SQLAlchemy for a widely used DB connection wrapper.  This plays
nicely with libraries like pandas.

How this addresses that need:
* Adds new AthenaClient, that will get composed with the
soon-to-be AIPValidator class

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IR-182