Add AIP class that encapsulates validation logic
Why these changes are being introduced:

The primary purpose of this AWS lambda is performing AIP validation.

This does not yet handle invocation of this new AIP.validate() with
the lambda getting called; this will be next steps.

How this addresses that need:
* A new AIP class is introduced that expects an S3 URI where the AIP
is stored, and provides a method validate() to validate the AIP
* This completes the bulk of porting over spike code to this
lambda repository

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IR-182