Wire AIP validator up to lambda handler
Why these changes are being introduced:

With the AIP class fully encapsulating AIP validation logic,
the next step was to wire this up to an AWS lambda handler.

This handler is prepared to handle invocations that may, or
may not, come from HTTP requests (e.g. Application Load Balancer (ALB)
or Function URL), which have varying forms for the passed 'event' object.

How this addresses that need:
* Completes the stubbed function 'validator.lambda_handler'
* Lambda handler handles parsing of input payload, validation of payload,
checking of a 'challenge secret' for some lightweight authorization, and
then invoking the AIP class to validate the AIP.

Side effects of this change:
* Lambda is callable in a deployed fashion.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IR-183