Get most recent partition of each glue table
Why these changes are being introduced:

An edge case was discovered where, if the query was run
during an S3 inventory data load, the value of 'max(dt)'
could be the value from a Glue table that had already been
updated, but others had not.  In this way, it would look
like -- for that dt partition -- that the files didn't exist.

How this addresses that need:

This simplifies the query by filtering each glue table
immediately to the most recent dt partition.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IR-193