Add CLI "sidecar" to application
Why these changes are being introduced:

This s3 bagit validator is deployed as a lambda for calling
via HTTP requests.  It was designed this way for flexibility as
to who, what, where, and when might invoke it.

Until more automation is established, that potentially wires
this lambda into Archivematica, it's quite possible that
a person will be manually validating bags (potentially in
bulk).   This CLI, as part of the application, provides that
approach.

How this addresses that need:
* A CLI is included in the repository, where the actual
work is performed by invoking the deployed lambda.

Side effects of this change:
* The lambda can be utilized, with all its horizontal scaling
and pre-configured permissions, via a CLI.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IR-191