Gracefully handle missing or empty S3 inventory location
Why these changes are being introduced:

If an S3 inventory prefix does not resolve to any files,
or the location exists but no Inventory data has been
written yet, it was throwing an opaque error.

How this addresses that need:

The solution is twofold:
1. issue a warning log that indicates the configured S3
inventory location either does not exist or is empty
2. return an empty list and continue

While it is unusual for a configured S3 inventory to not
have symlink.txt files, it is possible.  By issuing a
logging warning we see this in Sentry, and it's visible
in logs if we want to understand why an AIP won't validate.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IR-209