Bubble up handled exceptions into lambda response
Why these changes are being introduced:

It was pointed out that even for some handled exceptions (e.g.
a bad UUID) the error message returned by the Lambda + CLI was
still opaque.  Effort had gone into catching those cases and
providing meaningful python Exceptions, but they weren't getting
returned by the Lambda.

Turns out the CLI has some copy pasta for requests that was
raising an exception based on a HTTP status code, which was
short-circuiting the returning of those curated exceptions.

How this addresses that need:

Remove the .raise_for_status() call in the CLI which allows
the exceptions to get returned in the Lambda response.  Also,
improve exception handling to expose when a UUID is not found.

Side effects of this change:
* Lambda, and by proxy CLI, responses are improved with more
details when something fails

Relevant ticket(s):
* None