Modify defaults for response payload
Why these changes are being introduced:

It was discovered that a lambda behind an ALB has a max response size
of 1mb, which is fairly small.  Formerly, we were returning the full AIP
manifest in the response regardless of validation results, meaning we
might return a manifest with 8k files where all the checksums
were confirmed.

This amount of data returned from the Lambda was exceeding this 1mb
limit.

How this addresses that need:
* Add include/exclude filters on the ValidationResponse dataclass,
allowing control over what fields to return
* Removes 'manifest' from payload when serializing Lambda response
* Add 'error_details' property that anticipates returning only files
that failed validation, which theoretically is much smaller

Side effects of this change:
* Full AIP manifest is not returned for every AIP validation.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IR-220