Add error_detail to lambda response
Why these changes are being introduced:

After dialing back the verbosity of the response, namely removing
the manifest, we need slightly dial up the verbosity of what went
wrong (if anything).

Testing indicated that the Lambda / CLI responses were missing
important information about why an AIP may have failed valiation.

It is likely more grain will be coming for validation issues,
but this is a good step in that direction.

How this addresses that need:
* New 'error_details' object added to the repsonse in addition
to the high level 'error' string that was present already

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IR-220