Move checksum complexity to S3Client class
Why these changes are being introduced:

This is tangentially related to relaxing requirements of up-to-date
S3 Inventory data, and also the organization of functionality
and method naming.

Upon deciding that we should calculate a checksum even if
Inventory data claims an object has a checksum, but it does
not, it was clear that the logic complexity was building
in the AIP class when it felt like it could be moved to the
S3Client class.

How this addresses that need:
* Create a single entrypoint method on the S3Client class,
'get_object_checksum'
* This method handles _how_ to get that checksum, but always
provides one
* In the process, some minor renaming of methods now that
we are more interested in what they are doing vs why

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IR-222