Add AWS SAM CLI testing harness
Why these changes are being introduced:

This lambda application is a bit unique in that we include a CLI "sidecar"
that is designed to invoked the deployed AWS Lambda.  It is unknown how
long this CLI will be required (bit of a stopgap measure) but for as
long as it it is included, testing it is important.

The testing loop was made difficult when the CLI commands would hit
the deployed lambda, requiring quite a few builds + deploys to
Dev environment (or even worse, Stage if we needed assets there).

How this addresses that need:

AWS SAM (Serverless Application Model) CLI is a way to test AWS
lambdas quite thoroughly locally.  Of particular help and interest
to this project is the ability to handle HTTP requests to the
lambda as if it were handling them from behind an ALB, Function
URL, or API gateway, which is how this project's CLI invokes the
lambda.

By including some scaffolding and Makefile commands we can
fairly easily run the lambda locally via AWS SAM CLI, allowing
for easier testing of the CLI.

Side effects of this change:
* None really

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IR-228