Improve AIP Inventory SQL query
Why these changes are being introduced:

As noted Jira ticket IR-232, one "AIP" in S3 was problematic.
It was an "AIP" in the sense it had a valid UUID in an S3 key,
but this UUID showed up only for a single object which was a
.7z file.  This actually pointed at a larger issue, how we
determine "AIPs" that we have found in Inventory data.

What we lacked before was some kind of check that "AIPs" we
found from Inventory data aligned with a valid Bagit structure.

How this addresses that need:

The SQL query that looks for S3 objects with a valid UUID
in the key, after grouping by found UUID + prefix, now
also ensures that 'bagit.txt' was included for at least
one of the AIP files.  Without this, we can assume
whatever was found in S3 was not a valid Bagit AIP
structure.

Side effects of this change:
* The amount of AIPs "found" appears to be less, but is
arguably just more accurate.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IR-232