Normalize UUID vs URI behavior for init and response
Why these changes are being introduced:

The application evolved from expecting an S3 URI as input for
AIP validation to expecting an AIP UUID.  While both were technically
supported before this change, the response was inconsistent and bulk
validation behavior inconsistent.

This work normalizes the instantiation of an AIP class instance
and the ValidationResponse to treat UUIDs and URIs as relatively
identical as inputs.  This paves the way for improving the bulk
validation handling of previous results.

How this addresses that need:
* An AIP is usually instantiated from AIP.from_uuid() or AIP.from_uri()
to normalize that bucket, AIP UUID, and AIP S3 URI are always set.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IR-228