Handle AIP UUID or URI for bulk retry and resume
Why these changes are being introduced:

Formerly, the bulk-validation retry and resume was heavily geared
towards the input CSV using AIP UUIDs.  While bulk validation supports
either UUID or URI columns, the retry and resume functionality behaved
inconsistently depending on which was used.

Furthermore, the logic for reusing previous was arguably more complex
than it needed to be.

This work is made easier by a previous commit that normalizes
the handling of UUIDs and URIs in the AIP class.

How this addresses that need:
* Distinct blocks for with or without pre-existing results
* Build a dataframe of existing results, use this for logging
and explicit list of AIPs to skip
* Check for both UUIDs or URIs in the new input CSV compared
against existing UUIDs or URIs

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IR-228