Multi-Region Deployment Automation
Why these changes are being introduced:
To enable this application to validate AIPs in S3 buckets in multiple
regions, GHA needs to be able to build the container and push it to
multiple regions. Additionally, since Infra is always involved in
managing the GHA-based deployment workflows, they are now explicitly
listed in the CODEOWNERS file.

How this addresses that need:
* Update the three GHA workflows to include the extra lines generated
by the mitlib-tf-workloads-ecr repository to enable deployment to the
us-west-2 region
* Update the CODEOWNERS file to force Infra to approve PRs that modify
the deployment workflows
* Update the README to include the linked infrastructure repositories

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IR-238